+++ electron.vite.config.ts
@@ -39,7 +39,7 @@ export default defineConfig(({ mode }) => {
             /^mongodb/,
             /^mssql/,
             /^typeorm-aurora-data-api-driver/,
-            /^typeorm/
+            /^typeorm/,
           ],
         },
       },