+++ src/renderer/config/Route.tsx
@@ -1,16 +1,13 @@
-import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import { MainLayout } from '../layout/MainLayout';
-import { Home } from '../pages/Home/Home';
-import SettingsPage from '../pages/Settings/SettingPage';
-import { UpdateDialog } from '../pages/Update/UpdateDialog';
-import { AboutPage } from '../pages/About/AboutPage';
+import { MainLayout } from '@/renderer/layout/MainLayout';
+import { Home } from '@/renderer/pages/Home/Home';
+import { UpdateDialog } from '@/renderer/pages/Update/UpdateDialog';
+import { AboutPage } from '@/renderer/pages/About/AboutPage';
 
 export const AppRoutes = () => (
   <Routes>
     <Route path="/" element={<MainLayout />}>
       <Route index element={<Home />} />
-      <Route path="settings" element={<SettingsPage />} />
       <Route path="about" element={<AboutPage />} />
     </Route>
     <Route path="/update-dialog" element={<UpdateDialog />} />