+++ src/renderer/i18n/i18n.ts
@@ -1,6 +1,6 @@
 import i18next from 'i18next';
 import { initReactI18next } from 'react-i18next';
-import { enTranslations, zhTranslations } from '@/shared/locales';
+import { enTranslations, zhTranslations } from '../../shared/locales';
 
 // eslint-disable-next-line import/no-named-as-default-member
 i18next.use(initReactI18next).init({